requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
colorama==0.4.4
Django==3.1.4
django-autoslug==1.9.8
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.8
pdfkit==0.6.1
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytils==0.3
pytz==2020.4
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
wkhtmltopdf==0.2
